requirements.txt
django==5.0
djangorestframework
channels
redis
celery
django-celery-beat
psycopg2-binary
python-decouple
